% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocenters.R
\docType{data}
\name{geocenters}
\alias{geocenters}
\title{Metadata for map ploting}
\format{
A data table object with 11 rows and 4 variables:
\describe{
  \item{region}{A region, as stated in colnames or rownames in a distance matrix}
  \item{y}{y coordinates for a region center}
  \item{x}{x coordinates for a region center}
  \item{col}{Color to use in hexadecimal format}
}
}
\source{
\url{http://www.diamondse.info/}
}
\usage{
geocenters
}
\description{
A dataset contains metadata, used for map ploting.
}
\keyword{datasets}
