% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReMIXTURE.R
\name{ReMIXTURE}
\alias{ReMIXTURE}
\title{ReMIXTURE}
\value{
R6 class object
}
\description{
The class, which makes provides an easy interface for ReMIXTURE methods, described here: (add reference)
}
\details{
Methods for this class are $run, $plot_heatmap, $plot. They all are described in a separate sections
}
\section{Warning}{

Does nothing interesting.
}

\examples{
test <- ReMIXTURE(distamce matrix, info_table)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ReMIXTURE$new()}}
\item \href{#method-run}{\code{ReMIXTURE$run()}}
\item \href{#method-plot_heatmap}{\code{ReMIXTURE$plot_heatmap()}}
\item \href{#method-plot_maps}{\code{ReMIXTURE$plot_maps()}}
\item \href{#method-clone}{\code{ReMIXTURE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$new(distance_matrix, info_table = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{distance_matrix}}{- A matrix of distances between samples.}

\item{\code{info_table}}{- an object of class data.frame, which describes region coordinatess and color to use}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$run(iterations = 1000, resample = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_heatmap"></a>}}
\if{latex}{\out{\hypertarget{method-plot_heatmap}{}}}
\subsection{Method \code{plot_heatmap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_heatmap()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_maps"></a>}}
\if{latex}{\out{\hypertarget{method-plot_maps}{}}}
\subsection{Method \code{plot_maps()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$plot_maps()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReMIXTURE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
